requirements.txt
Flask>=2.0
requests>=2.25
beautifulsoup4>=4.9
python-dotenv>=0.15
Werkzeug>=2.0 
gunicorn>=20.0 # Added for production deployment, good practice
